# swagger.yaml
swagger: "2.0"
info:
  description: "The REST API of Production Manager"
  version: "1.0.0"
  title: "ProductionManager"
basePath: "/api/"
...
schemes:
- "http"
paths:
...
  /objects:
    get:
      tags:
      - "objects"
      summary: "gets a list of objects from TDM"
      parameters: 
      - name: "filter"
        in: "query"
        description: "the filter created before"
        required: false
        type: string
      produces: 
      - "application/json"
      responses:
        200:
          description: "the objects"
          schema:
            type: array
            items:
              $ref: '#/definitions/Object'
  /objects/{objectId}/image:
    get:
      tags:
      - "objects"
      summary: "gets the image of a object"
      parameters:
      - name: "objectId"
        in: "path"
        description: "ID of object to return"
        required: true
        type: "string"
      produces:
      - "image/svg+xml"
      - "image/png"
      responses:
        200:
          description: "the image"
...
  /objects/{objectId}/print:
    post:
      tags:
      - "objects"
      summary: "prints this object"
      parameters:
      - name: "objectId"
        in: "path"
        description: "ID of object to return"
        required: true
        type: "string"
      responses:
        200:
          description: "triggered print"
...
  /localobjects:
    post:
      tags: 
      - "local objects"
      summary: "uplaod a new object"
      consumes: 
      - "application/x-www-form-urlencoded"
      - "multipart/form-data"
      parameters: 
      - name: file
        description: "The ufp containing gcode and image"
        in: formData
        type: file
      - name: title
        in: formData
        type: string
      responses:
        201:
          description: object created
          headers:
            Location:
              type: string
              description: "the path to the created object"
    get:
      tags:
      - "local objects"
      summary: "gets a list of local objects"
      parameters: 
      - name: "materials"
        in: "query"
        description: "the materialIds to filter the results, none equals all"
        required: false
        type: array
        items:
          type: string
      - name: "machines"
        in: "query"
        description: "the machinetypeIds to filter the results, none equals all"
        required: false
        type: array
        items:
          type: string
      - name: lang
        in: "query"
        description: "defines the language the texts should be returned"
        required: true
        type: "string"
      produces: 
      - "application/json"
      responses:
        200:
          description: "the local objects"
          schema:
            type: array
            items:
              $ref: '#/definitions/LocalObject' 
  /localobjects/{localObjectId}:
    get:
      tags:
      - "local objects"
      summary: "gets a local objects"
      parameters: 
      - name: "localObjectId"
        in: "path"
        description: "ID of object to return"
        required: true
        type: "string"
      produces: 
      - "application/json"
      responses:
        200:
          description: "the local object"
          schema:
            $ref: '#/definitions/LocalObject' 
              
    delete:
      tags:
      - "local objects"
      summary: "delete a local objects"
      parameters: 
      - name: "localObjectId"
        in: "path"
        description: "ID of object to delete"
        required: true
        type: "string"
      produces: 
      - "application/json"
      responses:
        200:
          description: "deleted"

    patch:
      tags:
        - "local objects"
      summary: "can be used to update title or description of a local object"
      parameters:
      - name: "localObjectId"
        in: "path"
        description: "ID of object to delete"
        required: true
        type: "string"
      - name: patch
        in: "body"
        description: "the properties of the local object to patch"
        required: true
        schema:
         $ref: '#/definitions/LocalObjectPatch'
      produces:
      - "application/json"
      responses:
        200:
          description: "the local object"
          schema:
            $ref: '#/definitions/LocalObject'
  /localobjects/{localObjectId}/image:
    get:
      tags:
      - "local objects"
      summary: "gets the image of a local object"
      parameters:
      - name: "localObjectId"
        in: "path"
        description: "ID of object to return"
        required: true
        type: "string"
      produces:
      - "image/png"
      responses:
        200:
          description: "the image"
  /localobjects/{localObjectId}/publish:
    post:
      tags:
        - "local objects"
        - "objects"
      summary: "starts upload process to tdm"
      parameters:
      - name: "localObjectId"
        in: "path"
        description: "ID of object to return"
        required: true
        type: "string"
      - name: publishInfo
        in: "body"
        description: "the information for publish"
        required: true
        schema:
         $ref: '#/definitions/PublishInfo'
      responses:
        201:
          description: "published"
  /localobjects/{localObjectId}/publish/retry:
    post:
      tags:
        - "local objects"
        - "objects"
      summary: "retries upload process to tdm"
      parameters:
      - name: "localObjectId"
        in: "path"
        description: "ID of object to return"
        required: true
        type: "string"
      responses:
        200:
          description: "triggered retry"
  /localobjects/{localObjectId}/publish/reset:
    post:
      tags:
        - "local objects"
        - "objects"
      summary: "resets upload process to tdm"
      parameters:
      - name: "localObjectId"
        in: "path"
        description: "ID of object to return"
        required: true
        type: "string"
      responses:
        200:
          description: "triggered retry"
  /localobjects/{localObjectId}/print:
    post:
      tags:
        - "local objects"
      summary: "prints the local object"
      parameters:
      - name: "localObjectId"
        in: "path"
        description: "ID of object to return"
        required: true
        type: "string"
      - name: printInfo
        in: "body"
        description: "the information for printing"
        required: true
        schema:
         $ref: '#/definitions/PrintInfo'
      responses:
        200:
          description: "triggered print"
definitions:
  LocalObject:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      description:
        type: string
      createdAt:
        type: string
      materials:
        type: array
        items:
          type: string
      machines:
        type: array
        items:
          type: string
  Object:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      description:
        type: string
      licenseFee:
        type: integer
      productCode:
        type: integer
      backgroundColor:
        type: string
      materials:
        type: array
        items:
          $ref: '#/definitions/Component'
      machines:
        type: array
        items:
          $ref: '#/definitions/Component'
...
  Component:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      parentId:
        type: string
      id:
        type: string
      displayColor:
        type: string
      attributes:
        type: array
        items:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
...
  LocalObjectPatch:
    type: "object"
    properties:
      name:
        type: string
      description:
        type: string
  PublishInfo:
    type: "object"
    properties:
      name:
        type: string
      description:
        type: string
      licenseFee:
        type: integer
  PrintInfo:
    type: "object"
    properties:
      machineId:
        type: string
...